requirements.txt
adlfs==2024.7.0
azure-common==1.1.28
azure-core==1.31.0
azure-datalake-store==0.0.53
azure-identity==1.18.0
azure-mgmt-compute==33.0.0
azure-mgmt-core==1.4.0
azure-mgmt-monitor==6.0.2
azure-mgmt-network==26.0.0
azure-mgmt-resource==23.1.1
azure-storage-blob==12.23.0
fsspec==2024.9.0
ipykernel==6.29.5
ipython==8.12.3
jupyter-client==8.6.3
jupyter-core==5.7.2
msal==1.31.0
msal-extensions==1.2.0
numpy==1.24.4
pandas==2.0.3
python-dotenv==1.0.1
